毕业设计成品

+++ src/main/java/com/yongxv/bishe/zhanshi/service/impl/UserServiceImpl.java
@@ -7,11 +7,14 @@ package com.yongxv.bishe.zhanshi.service.impl;
  */
 
 import com.alibaba.fastjson.JSON;
+import com.yongxv.bishe.zhanshi.domain.Goods;
 import com.yongxv.bishe.zhanshi.entity.*;
 import com.yongxv.bishe.zhanshi.mapper.BisheMapper;
+import com.yongxv.bishe.zhanshi.mapper.StoreMapper;
 import com.yongxv.bishe.zhanshi.repository.GoodsRepository;
 import com.yongxv.bishe.zhanshi.service.UserService;
 import com.yongxv.bishe.zhanshi.utils.DateUtil;
+import com.yongxv.bishe.zhanshi.utils.DateUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.apache.commons.lang3.StringUtils;
@@ -30,9 +33,11 @@ public class UserServiceImpl implements UserService {
     private BisheMapper bisheMapper;
     @Autowired
     private GoodsRepository goodsRepository;
+    @Autowired
+    private StoreMapper storeMapper;
 
     @Override
-    public Response login(String account, String passWord){
+    public Response login(String account, String passWord,String role){
         if(StringUtils.isBlank(account) || StringUtils.isBlank(passWord)){
             return Response.error(1,"请输入账号或者密码");
         }else{
@@ -41,30 +46,11 @@ public class UserServiceImpl implements UserService {
                 return Response.error(1,"用户不存在");
             }else{
                 if(user.getPassword().equals(passWord)){
-                    UserDo userDo = new UserDo();
-                    long consumers = bisheMapper.selectConsumers();
-                    long store = bisheMapper.selectStore();
-                    long shoes = bisheMapper.selectShoes();
-                    long clothes = bisheMapper.selectClothes();
-                    long others = bisheMapper.selectOthers();
-                    long shoesGoods = goodsRepository.query("0");
-                    long clothesGoods = goodsRepository.query("1");
-                    long othersGoods = goodsRepository.query("2");
-                    userDo.setId(user.getId());
-                    userDo.setUserName(user.getUserName());
-                    userDo.setType(user.getType());
-                    userDo.setStoreNumber(store);
-                    userDo.setUserNumber(store+consumers);
-                    userDo.setConsumers(consumers);
-                    long goodsTotal = shoesGoods+clothesGoods+othersGoods;
-                    userDo.setGoods(goodsTotal);
-                    userDo.setShoes(efficient(store,shoes));
-                    userDo.setClothes(efficient(store,clothes));
-                    userDo.setOthers(efficient(store,others));
-                    userDo.setShoesGoods(efficient(goodsTotal,shoesGoods));
-                    userDo.setClothesGoods(efficient(goodsTotal,clothesGoods));
-                    userDo.setOthersGoods(efficient(goodsTotal,othersGoods));
-                    return Response.success(userDo,"登陆成功");
+                    if(role.equals(user.getType())){
+                        return Response.success(user,"登陆成功");
+                    }else {
+                        return Response.error(1, "角色选择错误请重新选择");
+                    }
                 }else {
                     return Response.error(1,"账号或密码错误");
                 }
@@ -79,10 +65,10 @@ public class UserServiceImpl implements UserService {
             return Response.error(1,"账号或密码不能为空");
         }
         if(account.length() < 6 || passWord.length() < 6){
-            return Response.error(1,"账号或和密码长度必需大于6位");
+            return Response.error(1,"账号或和密码长度必需大于5位");
         }
         if(account.length() >15 || passWord.length() > 15){
-            return Response.error(1,"账号或和密码长度必需小于15位");
+            return Response.error(1,"账号或和密码长度必需小于16位");
         }
         if(!isNumeric(account)){
             return Response.error(1,"账号只能由纯数字组成");
@@ -111,6 +97,38 @@ public class UserServiceImpl implements UserService {
 
     }
 
+    @Override
+    public Response addAdmin(String account,String passWord,String userName,String introduction,String area){
+        if(StringUtils.isBlank(account) || StringUtils.isBlank(passWord)){
+            return Response.error(1,"账号或密码不能为空");
+        }
+        if(account.length() < 6 || passWord.length() < 6){
+            return Response.error(1,"账号或和密码长度必需大于5位");
+        }
+        if(account.length() >15 || passWord.length() > 15){
+            return Response.error(1,"账号或和密码长度必需小于16位");
+        }
+        if(!isNumeric(account)){
+            return Response.error(1,"账号只能由纯数字组成");
+        }
+        User user = bisheMapper.selectUser(account);
+        if(user == null || "".equals(user)){
+            User user1 = new User();
+            String today = DateUtil.getDateTime(DateUtil.getTimePattern(), new Date());
+            user1.setAccount(account);
+            user1.setPassword(passWord);
+            user1.setUserName(userName);
+            user1.setStoreType("");
+            user1.setIntroduction(introduction);
+            user1.setArea(area);
+            user1.setCreateTime(today);
+            bisheMapper.addAdmin(user1);
+            return Response.success(0,"添加成功");
+        }else{
+            return Response.error(1,"此账号已存在");
+        }
+    }
+
 
     @Override
     public Response updateInformation(int id,String userName,String introduction,String area){
@@ -128,20 +146,23 @@ public class UserServiceImpl implements UserService {
     @Override
     public Response updatePassWord(int id,String account,String password,String passwordAgain){
         User user = bisheMapper.selectUserByid(id);
+        if(StringUtils.isBlank(password) || StringUtils.isBlank(passwordAgain)){
+            return Response.error(1,"密码不可为空");
+        }
         if(!user.getAccount().equals(account)){
             return Response.error(1,"账号输入不正确");
         }else if(!password.equals(passwordAgain)){
             return Response.error(1,"两次密码输入不相同");
         }else if(password.length() < 6){
-            return Response.error(1,"密码必须大于6位");
+            return Response.error(1,"密码必须大于5位");
         }else if(password.length() > 15){
-            return Response.error(1,"密码必须小于15位");
+            return Response.error(1,"密码必须小于16位");
         }else {
             User user1 = new User();
             user1.setId(id);
             user1.setPassword(password);
             bisheMapper.updatePassword(user1);
-            return Response.success("","修改成功请重新登陆");
+            return Response.success("","修改成功");
         }
 
     }
@@ -155,14 +176,32 @@ public class UserServiceImpl implements UserService {
     @Override
     public Response focus(int uid,int uuid){
         String today = DateUtil.getDateTime(DateUtil.getTimePattern(), new Date());
-        bisheMapper.addFocus(uid,uuid,today);
-        return Response.success(0,"关注成功");
+        User user  = bisheMapper.selectUserByid(uuid);
+        List<Associated> associatedList = new ArrayList<>();
+        associatedList = bisheMapper.selectGuaZhu(uid,uuid);
+        if(associatedList.size() != 0){
+            return Response.error(1,"你已进行关注");
+        }else{
+            bisheMapper.addFocus(uid,uuid,today);
+            bisheMapper.updateFocusCount((Integer.parseInt(user.getFocusCount())+1)+"",uuid);
+            return Response.success(1,"关注成功");
+        }
+
     }
 
     @Override
     public Response deleteFocus(int uid,int uuid){
-        bisheMapper.deleteFocus(uid,uuid);
-        return Response.success(0,"取消关注成功");
+        List<Associated> associatedList = new ArrayList<>();
+        associatedList = bisheMapper.selectGuaZhu(uid,uuid);
+        User user  = bisheMapper.selectUserByid(uuid);
+        if(associatedList.size() != 0){
+            bisheMapper.deleteFocus(uid,uuid);
+            bisheMapper.updateFocusCount((Integer.parseInt(user.getFocusCount())-1)+"",uuid);
+            return Response.success(0,"取消关注成功");
+        }else{
+            return Response.error(1,"未进行关注");
+        }
+
     }
 
     @Override
@@ -181,7 +220,7 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public Response selectFocusUser(int uuid,int page,int size){
-        List<Associated> associatedList = bisheMapper.selectFocusUser(uuid,page*size,size);
+        List<Associated> associatedList = bisheMapper.selectFocusUser(uuid,(page-1)*size,size);
         List<User> users = new ArrayList<>();
         for (int i = 0; i <associatedList.size() ; i++) {
             User user  = bisheMapper.selectUserByid(associatedList.get(i).getUid());
@@ -194,15 +233,26 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public Response selectNewUpdate(){
-        String today = DateUtil.getDate(new Date());
-        String startTime = today+" 00:00:00";
-        List<User> associatedList = bisheMapper.selectStoreByUpdatetime(startTime);
-        return Response.success(associatedList);
+    public Response selectNewUpdate(int page,int size){
+        String day = DateUtils.getDayAgoOrAfterString(-3);
+        List<UserDto> users = bisheMapper.selectStoreByUpdatetime(day,(page-1)*size,size);
+        Integer count = bisheMapper.selectStoreByUpdatetimeCount(day);
+        for (int i = 0; i < users.size(); i++) {
+            Goods goods = goodsRepository.queryOneByUid(users.get(i).getId() + "");
+            if(goods == null || "".equals(goods)) {
+                users.get(i).setPicture("");
+            }else {
+                users.get(i).setPicture(goods.getPicture());
+            }
+        }
+        Tatal tatal = new Tatal();
+        tatal.setTotal(count);
+        return Response.success(users,tatal);
     }
+
     @Override
     public Response selectAcess(int uid,int page,int size){
-        List<Access> accessList = bisheMapper.selectAccess(uid,page*size,size);
+        List<Access> accessList = bisheMapper.selectAccess(uid,(page-1)*size,size);
         Integer count = bisheMapper.selectAccessCount(uid);
         Tatal tatal = new Tatal();
         tatal.setTotal(count);
@@ -221,6 +271,19 @@ public class UserServiceImpl implements UserService {
         User user = bisheMapper.selectUserByid(id);
         if(user.getType().equals("2")){
             goodsRepository.delete(id+"");
+            storeMapper.deleteAccessDian(id);
+            storeMapper.deleteAssociatedDian(id);
+            storeMapper.deleteContentDian(id);
+        }
+        if(user.getType().equals("1")){
+            storeMapper.deleteAccessRen(id);
+            storeMapper.deleteContentRen(id);
+            List<Associated> associatedList = storeMapper.selectFocusAll(id);
+            for (int i = 0; i <associatedList.size() ; i++) {
+                User user1 = bisheMapper.selectUserByid(associatedList.get(i).getUuid());
+                bisheMapper.updateFocusCount((Integer.parseInt(user1.getFocusCount())-1)+"",associatedList.get(i).getUuid());
+            }
+            storeMapper.deleteAssociatedRen(id);
         }
         bisheMapper.deleteUser(id);
         return Response.success(0);
@@ -228,7 +291,7 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public Response selectStore(int page,int size){
-        List<User> userList = bisheMapper.selectStores(page*size,size);
+        List<User> userList = bisheMapper.selectStores((page-1)*size,size);
         Integer count = bisheMapper.selectStore();
         Tatal tatal = new Tatal();
         tatal.setTotal(count);
@@ -237,7 +300,7 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public Response selectConsumers(int page,int size){
-        List<User> userList = bisheMapper.selectConsumer(page*size,size);
+        List<User> userList = bisheMapper.selectConsumer((page-1)*size,size);
         Integer count = bisheMapper.selectConsumers();
         Tatal tatal = new Tatal();
         tatal.setTotal(count);
@@ -285,13 +348,13 @@ public class UserServiceImpl implements UserService {
     public Response receivedPushOrMessage(int id,int page,int size){
         User user = bisheMapper.selectUserByid(id);
         if(user.getType().equals("1")){
-            List<Content> contentList =  bisheMapper.receivedPush(id,page*size,size);
+            List<Content> contentList =  bisheMapper.receivedPush(id,(page-1)*size,size);
             Integer count = bisheMapper.receivedPushCount(id);
             Tatal tatal = new Tatal();
             tatal.setTotal(count);
             return Response.success(contentList,tatal);
         }else{
-            List<Content> contentList =  bisheMapper.receivedMessage(id,page*size,size);
+            List<Content> contentList =  bisheMapper.receivedMessage(id,(page-1)*size,size);
             Integer count = bisheMapper.receivedMessageCount(id);
             Tatal tatal = new Tatal();
             tatal.setTotal(count);
@@ -300,7 +363,7 @@ public class UserServiceImpl implements UserService {
     }
 
 
-
+    @Override
     public Response deletePushAndMessage(int id,int uid){
         User user = bisheMapper.selectUserByid(uid);
         if(user.getType().equals("1")){
@@ -310,6 +373,22 @@ public class UserServiceImpl implements UserService {
         }
         return Response.success(0);
     }
+
+    @Override
+    public Response selectUserBypage(int id,int page,int size){
+        List<User> list = bisheMapper.selectUserByPage(id,(page-1)*size,size);
+        Integer count = bisheMapper.selectUserCount(id);
+        Tatal tatal = new Tatal();
+        tatal.setTotal(count);
+        return Response.success(list,tatal);
+    }
+
+    @Override
+    public Response deleteAccess(int id){
+        bisheMapper.deleteAccess(id);
+        return Response.success(0,"删除访问记录成功");
+    }
+
 
 
     /**
